+++ server-new/index.ts
@@ -1,4 +1,4 @@
-import { log } from '~/common/log';
+import { log } from '~/utils/log';
 import { setupServer } from '~/setup/setup';
 import { initSentry } from '~/setup/sentry';
 import { server } from '~/setup/server';
